ad service for fetchEntrypointGroups
