+++ Projects/App/Sources/AppDelegate.swift
@@ -305,6 +305,7 @@ extension ApolloClient {
                 let networkClient = NetworkClient()
                 let messagesClient = FetchMessagesClientOctopus()
                 let sendMessage = SendMessagesClientOctopus()
+                let hFetchEntrypointsService = FetchEntrypointsServiceOctopus()
                 Dependencies.shared.add(module: Module { hApollo.octopus })
                 Dependencies.shared.add(module: Module { () -> ChatFileUploaderClient in networkClient })
                 Dependencies.shared.add(module: Module { () -> FetchMessagesClient in messagesClient })
@@ -326,6 +327,9 @@ extension ApolloClient {
                     Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
                     Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
                     Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
+                    Dependencies.shared.add(
+                        module: Module { () -> hFetchEntrypointsService in hFetchEntrypointsService }
+                    )
                 case .production, .custom:
                     let hFetchClaimService = FetchClaimServiceOctopus()
                     Dependencies.shared.add(module: Module { () -> FileUploaderClient in networkClient })
@@ -335,7 +339,9 @@ extension ApolloClient {
                     Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
                     Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
                     Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
-
+                    Dependencies.shared.add(
+                        module: Module { () -> hFetchEntrypointsService in hFetchEntrypointsService }
+                    )
                 }
             }
             .toVoid()
