+++ Projects/Claims/Sources/Services/DemoImplementation/FetchEntrypointsDemoClient.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+public class FetchEntrypointsServiceDemo: hFetchEntrypointsService {
+    public init() {}
+    public func get() async throws -> [ClaimEntryPointGroupResponseModel] {
+        return [
+            ClaimEntryPointGroupResponseModel(
+                id: "entrypointId",
+                displayName: "EntrypointGroup",
+                entrypoints: []
+            )
+        ]
+    }
+}
