+++ Projects/Claims/Sources/Services/DemoImplementation/UploadClaimFileDemoClient.swift
@@ -1,0 +1,12 @@
+import Foundation
+import hCore
+
+public class hClaimFileUploadServiceDemo {
+    func upload(
+        endPoint: String,
+        files: [File],
+        withProgress: ((_ progress: Double) -> Void)?
+    ) async throws -> [ClaimFileUploadResponse] {
+        return []
+    }
+}
