+++ Projects/Claims/Sources/Services/OctopusImplementation/UploadClaimFileServiceOctopus.swift
@@ -5,24 +5,6 @@ import UIKit
 import hCore
 import hGraphQL
 
-public protocol hClaimFileUploadService {
-    func upload(
-        endPoint: String,
-        files: [File],
-        withProgress: ((_ progress: Double) -> Void)?
-    ) async throws -> [ClaimFileUploadResponse]
-}
-
-public class hClaimFileUploadServiceDemo {
-    func upload(
-        endPoint: String,
-        files: [File],
-        withProgress: ((_ progress: Double) -> Void)?
-    ) async throws -> [ClaimFileUploadResponse] {
-        return []
-    }
-}
-
 extension NetworkClient: hClaimFileUploadService {
     public func upload(
         endPoint: String,
