+++ Projects/Claims/Sources/Services/Protocols/FetchClaimService.swift
@@ -1,0 +1,7 @@
+import Foundation
+import hCore
+
+public protocol hFetchClaimService {
+    func get() async throws -> [ClaimModel]
+    func getFiles() async throws -> [String: [File]]
+}
