+++ Projects/Claims/Sources/Services/Protocols/FetchEntrypointsService.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol hFetchEntrypointsService {
+    func get() async throws -> [ClaimEntryPointGroupResponseModel]
+}
