+++ Projects/Claims/Sources/Services/Protocols/UploadClaimFileService.swift
@@ -1,0 +1,10 @@
+import Foundation
+import hCore
+
+public protocol hClaimFileUploadService {
+    func upload(
+        endPoint: String,
+        files: [File],
+        withProgress: ((_ progress: Double) -> Void)?
+    ) async throws -> [ClaimFileUploadResponse]
+}
